refactor(Tooltip): client境界が必要なモジュールをclient/配下に移す
CLAUDE.mdのclient/ディレクトリ規約に合わせ、Tooltip・TooltipPortalを
client/components/Tooltip/配下へ移動する。

client/index.tsは作らず、Tooltip.tsxのみが'use client'を持つ構成のため、
バレルの平坦化による副作用importの混入は発生しない。
ビルド出力でTooltip.jsが境界として保持されることを確認済み。
